# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Stats {
  totalCategories: number;
  totalAudioStories: number;
  totalVideos: number;
  publishedAudioStories: number;
  publishedVideos: number;
}

export interface Category {
  id: number;
  name: string;
  label: string;
  icon: string;
  type: string;
  sortOrder: number;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryInput {
  name: string;
  label: string;
  icon: string;
  type: string;
  sortOrder?: number;
  active?: boolean;
}

export interface CategoryUpdate {
  name?: string;
  label?: string;
  icon?: string;
  type?: string;
  sortOrder?: number;
  active?: boolean;
}

export interface AudioStory {
  id: number;
  title: string;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  narrator: string;
  durationSeconds: number;
  description: string;
  /** @nullable */
  thumbnailUrl: string | null;
  audioUrl: string;
  sourceType: string;
  published: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface AudioStoryInput {
  title: string;
  /** @nullable */
  categoryId?: number | null;
  narrator: string;
  durationSeconds: number;
  description: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  audioUrl: string;
  sourceType: string;
  published?: boolean;
  sortOrder?: number;
}

export interface AudioStoryUpdate {
  title?: string;
  /** @nullable */
  categoryId?: number | null;
  narrator?: string;
  durationSeconds?: number;
  description?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  audioUrl?: string;
  sourceType?: string;
  published?: boolean;
  sortOrder?: number;
}

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  description: string;
  /** @nullable */
  thumbnailUrl: string | null;
  videoUrl: string;
  sourceType: string;
  /** @nullable */
  youtubeId: string | null;
  views: number;
  published: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface VideoInput {
  title: string;
  /** @nullable */
  categoryId?: number | null;
  description: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  videoUrl: string;
  sourceType: string;
  /** @nullable */
  youtubeId?: string | null;
  views?: number;
  published?: boolean;
  sortOrder?: number;
}

export interface VideoUpdate {
  title?: string;
  /** @nullable */
  categoryId?: number | null;
  description?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  videoUrl?: string;
  sourceType?: string;
  /** @nullable */
  youtubeId?: string | null;
  views?: number;
  published?: boolean;
  sortOrder?: number;
}

export interface YoutubeInfoRequest {
  url: string;
}

export interface YoutubeInfo {
  title: string;
  thumbnailUrl: string;
  youtubeId: string;
  description: string;
  duration: number;
  authorName?: string;
}

export interface UploadUrlRequest {
  /**
   * Original file name.
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes.
   * @minimum 1
   */
  size: number;
  /**
   * MIME type of the file (e.g. `image/jpeg`).
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListCategoriesParams = {
  type?: string;
};

export type ListAudioStoriesParams = {
  category?: string;
  published?: boolean;
};

export type ListVideosParams = {
  category?: string;
  published?: boolean;
};