# lib/api-zod/src/generated/types/audioStoryUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface AudioStoryUpdate {
  title?: string;
  /** @nullable */
  categoryId?: number | null;
  narrator?: string;
  durationSeconds?: number;
  description?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  audioUrl?: string;
  sourceType?: string;
  published?: boolean;
  sortOrder?: number;
}