# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface CategoryUpdate {
  name?: string;
  label?: string;
  icon?: string;
  type?: string;
  sortOrder?: number;
  active?: boolean;
}